Renamed and patch for password generator
